bpo-45193: Restore IDLE completion boxes on Ubuntu (GH-28343)
The line that should not have been needed on macOS tk 8.6.8 but was,
should not be a problem on Ubuntu, but is.  It is not needed on macOS
tk 8.6.11, installed with 3.10.  Disable it but leave it for
now in case some system needs it.
(cherry picked from commit 4557576a0053690d229098a29b55765ac08025f9)

Co-authored-by: Terry Jan Reedy <tjreedy@udel.edu>